requirements.txt
flask
flask-login
flask-sqlalchemy
flask-dance
python-dotenv
